+++ NEWS.md
@@ -1,3 +1,15 @@
+# jrt 1.1.0
+
+* Fixes for compatibility with recent versions of the `mirt` package.
+* Fixes for compatibility with recent versions of the `ggplot2` package.
+* Update package citation.
+* Added documentation of S4 classes/methods.
+* Added the possibility to override the automatic facet naming in `jcc.plot()` with argument `facet.names`.
+* Added the possibility to change the opacity of the different category lines in `jcc.plot()` with `line.opacity`.
+* Simplified messages for situations with judges/items with unobserved categories.
+* Made Akaike Information Criterion (AIC) the default option in model comparisons.
+
+
 # jrt 1.0.1
 
 * Simulated dataset is now directly included as `jrt::ratings`.